+++ adyen/pages/checkout/confirmation.jsx
