+++ overrides/app/routes.jsx
@@ -4,40 +4,14 @@
  * SPDX-License-Identifier: BSD-3-Clause
  * For full license text, see the LICENSE file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-
-import React from 'react'
-import loadable from '@loadable/component'
 import {getConfig} from '@salesforce/pwa-kit-runtime/utils/ssr-config'
 
 // Components
-import {Skeleton} from '@chakra-ui/react'
 import {configureRoutes} from '@salesforce/retail-react-app/app/utils/routes-utils'
 import {routes as _routes} from '@salesforce/retail-react-app/app/routes'
-import AdyenCheckoutError from '../../adyen/components/AdyenCheckoutError'
+import _adyenRoutes from '../../adyen/pages/routes'
 
-const fallback = <Skeleton height="75vh" width="100%" />
-
-// Create your pages here and add them to the routes array
-// Use loadable to split code into smaller js chunks
-const Checkout = loadable(() => import('./pages/checkout'), {fallback})
-const CheckoutConfirmation = loadable(() => import('./pages/checkout/confirmation'), {fallback})
-
-const routes = [
-    {
-        path: '/checkout',
-        component: Checkout,
-        exact: true
-    },
-    {
-        path: '/checkout/error',
-        component: AdyenCheckoutError
-    },
-    {
-        path: '/checkout/confirmation/:orderNo',
-        component: CheckoutConfirmation
-    },
-    ..._routes
-]
+const routes = [..._adyenRoutes, ..._routes]
 
 export default () => {
     const config = getConfig()
